+++ CHANGELOG.md
@@ -7,5 +7,6 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 
 ### Added
 
-- Added git hooks to ensure documentation and changelog are always up to date
-- Initial version
+- git hooks to ensure documentation and changelog are always up to date
+- clean output directory before tests
+- generates either SVG or PNG format